fix: make ids of type integer instead of long

+++ src/main/kotlin/edu/spring/stories/entities/Developer.kt
@@ -9,7 +9,7 @@ import jakarta.persistence.PreRemove
 class Developer {
 
     @Id
-    var id: Long? = null
+    var id: Int? = null
     var firstname: String? = null
     var lastname: String? = null
